+++ TMT/TMT/Map/Views/EndStopView.swift
@@ -1,68 +1,77 @@
-////
-////  EndStopView.swift
-////  TMT
-////
-////  Created by Choi Minkyeong on 11/1/24.
-////
 //
-//import SwiftUI
+//  EndStopView.swift
+//  TMT
 //
-//struct EndStopView: View {
-//    var endStop: String
-//    var remainingStops: Int
-//    
-//    var body: some View {
-//        let colors = getMainColor(remainingStops: remainingStops)
-//        ZStack {
-//            RoundedRectangle(cornerRadius: 16)
-//                .fill(colors.backgroundColor)
-//                .frame(height: 65)
-//                .shadow(color: .black.opacity(0.25), radius: 2.5, x: 0, y: 2)
-//            HStack {
-//                VStack(alignment: .leading, spacing: 0) {
-//                    HStack {
-//                        Image("\(colors.pinImage)")
-//                            .resizable()
-//                            .frame(width: 12, height: 12)
-//                        Text("Destination")
-//                            .foregroundStyle(colors.subTextColor)
-//                    }
-//                    Text("\(endStop)")
-//                        .lineLimit(1)
-//                        .truncationMode(.tail)
-//                }
-//                .padding(.leading, 24)
-//                .padding(.vertical, 12)
-//                
-//                Spacer(minLength: 24)
-//                
-//                HStack {
-//                    Text("\(remainingStops)")
-//                        .font(.title)
-//                        .bold()
-//                    VStack {
-//                        Spacer()
-//                        Text("left")
-//                    }
-//                }
-//                .frame(height: 18)
-//                .padding(.trailing, 24)
-//            }
-//            .foregroundStyle(colors.textColor)
-//            .cornerRadius(16)
-//        }
-//        .padding(.bottom, 18)
-//        .padding(.leading, 16.25)
-//        .padding(.trailing, 16.75)
-//    }
-//    
-//    /// EndStopView의 메인 컬러를 판단합니다.
-//    private func getMainColor(remainingStops: Int) -> (backgroundColor: Color, textColor: Color, subTextColor: Color, pinImage: String) {
-//        let status = StopStatusEnum(remainingStops: remainingStops)
-//        return (status.backgroundColor, status.textColor, status.subTextColor, status.pinImage)
-//    }
-//}
+//  Created by Choi Minkyeong on 11/1/24.
 //
-//#Preview {
-//    EndStopView(endStop: "Youngildae Beach", remainingStops: 20)
-//}
+
+import SwiftUI
+
+struct EndStopView: View {
+    var endStopNameKorean: String
+    var endStopNameRomanized: String
+    var endStopNameNaver: String
+    var remainingStops: Int
+    
+    var body: some View {
+        let colors = mainColor(remainingStops: remainingStops)
+        VStack(spacing: 0) {
+            Text("Destination")
+                .foregroundStyle(colors.destinationColor)
+                .frame(width: 355, height: 40.2)
+                .offset(y: -10)
+                .background {
+                    RoundedRectangle(cornerRadius: 20)
+                        .trim(from: 0.5, to: 1)
+                        .foregroundStyle(colors.backgroundColor)
+                }
+            
+            VStack(alignment: .leading, spacing: 0) {
+                Text("Korean")
+                    .foregroundStyle(.grey200)
+                    .font(.footnote)
+                    .padding(.top, 5.34)
+                Text("\(endStopNameKorean)")
+                    .foregroundStyle(.yellow900)
+                    .font(.title2)
+                    .fontWeight(.bold)
+                Text("[\(endStopNameRomanized)]")
+                    .foregroundStyle(.yellow900)
+                    .font(.footnote)
+                Divider()
+                    .foregroundStyle(.grey100)
+                    .padding(.vertical, 9)
+                Text("English")
+                    .foregroundStyle(.grey200)
+                    .font(.footnote)
+                Text("\(endStopNameNaver)")
+                    .foregroundStyle(.yellow900)
+                    .font(.title2)
+                    .fontWeight(.bold)
+            }
+            .padding(.leading, 16)
+            .padding(.bottom, 11.48)
+            .frame(width: 355, height: 145)
+            .background {
+                UnevenRoundedRectangle(cornerRadii: .init(
+                    topLeading: 0,
+                    bottomLeading: 15,
+                    bottomTrailing: 15,
+                    topTrailing: 0
+                ))
+                .foregroundStyle(.basicWhite)
+            }
+            .offset(y: -20.1)
+        }
+    }
+    
+    /// EndStopView의 메인 컬러를 판단합니다.
+    private func mainColor(remainingStops: Int) -> (backgroundColor: Color, leftStopNumberColor: Color, leftStopTextColor: Color, destinationColor: Color, pinImage: String) {
+        let status = StopStatusEnum(remainingStops: remainingStops)
+        return (status.backgroundColor, status.leftStopNumberColor, status.leftStopTextColor, status.destinationColor, status.pinImage)
+    }
+}
+
+#Preview {
+    EndStopView(endStopNameKorean: "포항제철고등학교", endStopNameRomanized: "[Pohang Je-cheol Go-deung-hak-gyo", endStopNameNaver: "Pohang Jecheol High School", remainingStops: 4)
+}
